requirements.txt
aiofiles==24.1.0
aiogram==3.17.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
asyncio==3.4.3
attrs==24.2.0
beautifulsoup4==4.12.3
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.8
colorama==0.4.6
distro==1.9.0
et_xmlfile==2.0.0
frozenlist==1.5.0
greenlet==3.1.1
gTTS==2.5.4
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jiter==0.8.2
lxml==5.3.0
magic-filter==1.0.12
multidict==6.1.0
numpy==2.1.3
openai==1.59.3
openpyxl==3.1.5
outcome==1.3.0.post0
pandas==2.2.3
propcache==0.2.1
psgfiglet==5.0.0
psycopg2==2.9.10
psycopg2-binary==2.9.10
pyasn1==0.6.0
pycparser==2.22
pydantic==2.10.4
pydantic_core==2.27.2
pyfiglet==1.0.2
PySimpleGUI==5.0.6
PySocks==1.7.1
pyTelegramBotAPI==4.26.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-telegram-bot==21.10
pytz==2024.2
requests==2.32.3
rsa==4.9
selenium==4.26.1
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
SQLAlchemy==2.0.36
tqdm==4.67.1
trio==0.27.0
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
websocket-client==1.8.0
wsproto==1.2.0
yarl==1.18.3
